画面上の全ての文言を多言語化 削除する写真が存在しない場合の例外処理を追加

+++ src/main/java/jp/yoshida/photoadmin/controller/form/validation/MaxFileSizeValidator.java
@@ -9,7 +9,7 @@ import java.util.Objects;
 
 public class MaxFileSizeValidator implements ConstraintValidator<MaxFileSize, MultipartFile> {
 
-    private int value;
+    private long value;
 
     private KeyWordsConstants.UNIT_FILE_SIZE unit;
